+++ src/stream/src/executor/top_n/top_n_plain.rs
@@ -206,1122 +206,3 @@ where
         None
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use assert_matches::assert_matches;
-    use futures::StreamExt;
-    use risingwave_common::array::stream_chunk::StreamChunkTestExt;
-    use risingwave_common::catalog::{Field, Schema};
-    use risingwave_common::types::DataType;
-    use risingwave_common::util::sort_util::OrderType;
-
-    use super::*;
-    use crate::executor::test_utils::top_n_executor::create_in_memory_state_table;
-    use crate::executor::test_utils::MockSource;
-    use crate::executor::{Barrier, Message};
-
-    mod test1 {
-        use super::*;
-        use crate::executor::ActorContext;
-        fn create_stream_chunks() -> Vec<StreamChunk> {
-            let chunk1 = StreamChunk::from_pretty(
-                "  I I
-                +  1 0
-                +  2 1
-                +  3 2
-                + 10 3
-                +  9 4
-                +  8 5",
-            );
-            let chunk2 = StreamChunk::from_pretty(
-                "  I I
-                +  7 6
-                -  3 2
-                -  1 0
-                +  5 7
-                -  2 1
-                + 11 8",
-            );
-            let chunk3 = StreamChunk::from_pretty(
-                "  I  I
-                +  6  9
-                + 12 10
-                + 13 11
-                + 14 12",
-            );
-            let chunk4 = StreamChunk::from_pretty(
-                "  I  I
-                -  5  7
-                -  6  9
-                - 11  8",
-            );
-            vec![chunk1, chunk2, chunk3, chunk4]
-        }
-
-        fn create_schema() -> Schema {
-            Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            }
-        }
-
-        fn storage_key() -> Vec<ColumnOrder> {
-            let mut v = order_by();
-            v.extend([ColumnOrder::new(1, OrderType::ascending())]);
-            v
-        }
-
-        fn order_by() -> Vec<ColumnOrder> {
-            vec![ColumnOrder::new(0, OrderType::ascending())]
-        }
-
-        fn pk_indices() -> PkIndices {
-            vec![0, 1]
-        }
-
-        fn create_source() -> Box<MockSource> {
-            let mut chunks = create_stream_chunks();
-            let schema = create_schema();
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(1)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Barrier(Barrier::new_test_barrier(2)),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Barrier(Barrier::new_test_barrier(3)),
-                    Message::Chunk(std::mem::take(&mut chunks[2])),
-                    Message::Barrier(Barrier::new_test_barrier(4)),
-                    Message::Chunk(std::mem::take(&mut chunks[3])),
-                    Message::Barrier(Barrier::new_test_barrier(5)),
-                ],
-            ))
-        }
-
-        #[tokio::test]
-        async fn test_top_n_executor_with_offset() {
-            let source = create_source();
-            let state_table = create_in_memory_state_table(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, false>::new(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (3, 1000),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    + 10 3
-                    +  9 4
-                    +  8 5"
-                )
-            );
-            // Barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  7 6
-                    -  7 6
-                    -  8 5
-                    +  8 5
-                    -  8 5
-                    + 11 8"
-                )
-            );
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            // (8, 9, 10, 11, 12, 13, 14)
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I  I
-                    +  8  5
-                    + 12 10
-                    + 13 11
-                    + 14 12"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            // (10, 12, 13, 14)
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  8 5
-                    -  9 4
-                    - 11 8"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-
-        #[tokio::test]
-        async fn test_top_n_executor_with_limit() {
-            let source = create_source();
-            let state_table = create_in_memory_state_table(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, false>::new(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (0, 4),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  1 0
-                    +  2 1
-                    +  3 2
-                    + 10 3
-                    - 10 3
-                    +  9 4
-                    -  9 4
-                    +  8 5"
-                )
-            );
-            // now () -> (1, 2, 3, 8)
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  8 5
-                    +  7 6
-                    -  3 2
-                    +  8 5
-                    -  1 0
-                    +  9 4
-                    -  9 4
-                    +  5 7
-                    -  2 1
-                    +  9 4"
-                )
-            );
-
-            // (5, 7, 8, 9)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  9 4
-                    +  6 9"
-                )
-            );
-            // (5, 6, 7, 8)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  5 7
-                    +  9 4
-                    -  6 9
-                    + 10 3"
-                )
-            );
-            // (7, 8, 9, 10)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-
-        // Should have the same result as above, since there are no duplicate sort keys.
-        #[tokio::test]
-        async fn test_top_n_executor_with_limit_with_ties() {
-            let source = create_source();
-            let state_table = create_in_memory_state_table(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, true>::new(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (0, 4),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  1 0
-                    +  2 1
-                    +  3 2
-                    + 10 3
-                    - 10 3
-                    +  9 4
-                    -  9 4
-                    +  8 5"
-                )
-            );
-            // now () -> (1, 2, 3, 8)
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 8 5
-                    + 7 6
-                    - 3 2
-                    + 8 5
-                    - 1 0
-                    + 9 4
-                    - 9 4
-                    + 5 7
-                    - 2 1
-                    + 9 4"
-                )
-            );
-
-            // (5, 7, 8, 9)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  9 4
-                    +  6 9"
-                )
-            );
-            // (5, 6, 7, 8)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    -  5 7
-                    +  9 4
-                    -  6 9
-                    + 10 3"
-                )
-            );
-            // (7, 8, 9, 10)
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-
-        #[tokio::test]
-        async fn test_top_n_executor_with_offset_and_limit() {
-            let source = create_source();
-            let state_table = create_in_memory_state_table(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, false>::new(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (3, 4),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    + 10 3
-                    +  9 4
-                    +  8 5"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  7 6
-                    -  7 6
-                    -  8 5
-                    +  8 5
-                    -  8 5
-                    + 11 8"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I I
-                +  8 5"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    "  I  I
-                    -  8  5
-                    + 12 10
-                    -  9  4
-                    + 13 11
-                    - 11  8
-                    + 14 12"
-                )
-            );
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-    }
-
-    mod test2 {
-        use risingwave_storage::memory::MemoryStateStore;
-
-        use super::*;
-        use crate::executor::test_utils::snapshot::check_until_pending;
-        use crate::executor::test_utils::top_n_executor::create_in_memory_state_table_from_state_store;
-        use crate::executor::ActorContext;
-
-        fn create_source_new() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 1 4 1001",
-                ),
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  5 1 4 1002 ",
-                ),
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 9 1 1003
-                +  9 8 1 1004
-                +  0 2 3 1005",
-                ),
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 0 2 1006",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(1)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Chunk(std::mem::take(&mut chunks[2])),
-                    Message::Chunk(std::mem::take(&mut chunks[3])),
-                    Message::Barrier(Barrier::new_test_barrier(2)),
-                ],
-            ))
-        }
-
-        fn create_source_new_before_recovery() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 1 4 1001",
-                ),
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  5 1 4 1002 ",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(1)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Barrier(Barrier::new_test_barrier(2)),
-                ],
-            ))
-        }
-
-        fn create_source_new_after_recovery() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 9 1 1003
-                +  9 8 1 1004
-                +  0 2 3 1005",
-                ),
-                StreamChunk::from_pretty(
-                    " I I I I
-                +  1 0 2 1006",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(3)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Barrier(Barrier::new_test_barrier(4)),
-                ],
-            ))
-        }
-
-        fn storage_key() -> Vec<ColumnOrder> {
-            order_by()
-        }
-
-        fn order_by() -> Vec<ColumnOrder> {
-            vec![
-                ColumnOrder::new(0, OrderType::ascending()),
-                ColumnOrder::new(3, OrderType::ascending()),
-            ]
-        }
-
-        fn pk_indices() -> PkIndices {
-            vec![0, 3]
-        }
-
-        #[tokio::test]
-        async fn test_top_n_executor_with_offset_and_limit_new() {
-            let source = create_source_new();
-            let state_table = create_in_memory_state_table(
-                &[
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                ],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, false>::new(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (1, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            check_until_pending(
-                &mut top_n_executor,
-                expect_test::expect![[r#"
-                    - !barrier 1
-                    - !chunk (empty)
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | + | 5 | 1 | 4 | 1002 |
-                      +---+---+---+---+------+
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | + | 1 | 9 | 1 | 1003 |
-                      | + | 9 | 8 | 1 | 1004 |
-                      | - | 9 | 8 | 1 | 1004 |
-                      | + | 1 | 1 | 4 | 1001 |
-                      +---+---+---+---+------+
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | - | 5 | 1 | 4 | 1002 |
-                      | + | 1 | 0 | 2 | 1006 |
-                      +---+---+---+---+------+
-                    - !barrier 2
-                    - !barrier 7
-                "#]],
-            )
-            .await;
-        }
-
-        #[tokio::test]
-        async fn test_top_n_executor_with_offset_and_limit_new_after_recovery() {
-            let state_store = MemoryStateStore::new();
-            let state_table = create_in_memory_state_table_from_state_store(
-                &[
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                ],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-                state_store.clone(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::<_, false>::new(
-                    create_source_new_before_recovery() as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (1, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            check_until_pending(
-                &mut top_n_executor,
-                expect_test::expect![[r#"
-                    - !barrier 1
-                    - !chunk (empty)
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | + | 5 | 1 | 4 | 1002 |
-                      +---+---+---+---+------+
-                    - !barrier 2
-                    - !barrier 5
-                "#]],
-            )
-            .await;
-
-            let state_table = create_in_memory_state_table_from_state_store(
-                &[
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                    DataType::Int64,
-                ],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-                state_store,
-            )
-            .await;
-
-            // recovery
-            let top_n_executor_after_recovery = Box::new(
-                TopNExecutor::<_, false>::new(
-                    create_source_new_after_recovery() as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (1, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor_after_recovery.execute();
-
-            check_until_pending(
-                &mut top_n_executor,
-                expect_test::expect![[r#"
-                    - !barrier 3
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | + | 1 | 9 | 1 | 1003 |
-                      | + | 9 | 8 | 1 | 1004 |
-                      | - | 9 | 8 | 1 | 1004 |
-                      | + | 1 | 1 | 4 | 1001 |
-                      +---+---+---+---+------+
-                    - !chunk |-
-                      +---+---+---+---+------+
-                      | - | 5 | 1 | 4 | 1002 |
-                      | + | 1 | 0 | 2 | 1006 |
-                      +---+---+---+---+------+
-                    - !barrier 4
-                    - !barrier 5
-                "#]],
-            )
-            .await;
-        }
-    }
-
-    mod test_with_ties {
-        use risingwave_storage::memory::MemoryStateStore;
-
-        use super::*;
-        use crate::executor::test_utils::top_n_executor::create_in_memory_state_table_from_state_store;
-        use crate::executor::ActorContext;
-
-        fn create_source() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  1 0
-                    +  2 1
-                    +  3 2
-                    + 10 3
-                    +  9 4
-                    +  8 5
-                    ",
-                ),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  3 6
-                    +  3 7
-                    +  1 8
-                    +  2 9
-                    + 10 10",
-                ),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 0",
-                ),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 8",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(1)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Chunk(std::mem::take(&mut chunks[2])),
-                    Message::Chunk(std::mem::take(&mut chunks[3])),
-                    Message::Barrier(Barrier::new_test_barrier(2)),
-                ],
-            ))
-        }
-
-        fn storage_key() -> Vec<ColumnOrder> {
-            let mut v = order_by();
-            v.push(ColumnOrder::new(1, OrderType::ascending()));
-            v
-        }
-
-        fn order_by() -> Vec<ColumnOrder> {
-            vec![ColumnOrder::new(0, OrderType::ascending())]
-        }
-
-        fn pk_indices() -> PkIndices {
-            vec![0, 1]
-        }
-
-        #[tokio::test]
-        async fn test_with_ties() {
-            let source = create_source();
-            let state_table = create_in_memory_state_table(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::new_with_ties_for_test(
-                    source as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (0, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    + 1 0
-                    + 2 1
-                    + 3 2"
-                )
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    + 3 6
-                    + 3 7
-                    - 3 7
-                    - 3 6
-                    - 3 2
-                    + 1 8
-                    + 2 9"
-                )
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 0"
-                )
-            );
-
-            // High cache has only 2 capacity, but we need to trigger 3 inserts here!
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 8
-                    + 3 2
-                    + 3 6
-                    + 3 7
-                    "
-                )
-            );
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-
-        fn create_source_before_recovery() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  1 0
-                    +  2 1
-                    +  3 2
-                    + 10 3
-                    +  9 4
-                    +  8 5
-                    ",
-                ),
-                StreamChunk::from_pretty(
-                    "  I I
-                    +  3 6
-                    +  3 7
-                    +  1 8
-                    +  2 9
-                    + 10 10",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(1)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Barrier(Barrier::new_test_barrier(2)),
-                ],
-            ))
-        }
-
-        fn create_source_after_recovery() -> Box<MockSource> {
-            let mut chunks = vec![
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 0",
-                ),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 8",
-                ),
-            ];
-            let schema = Schema {
-                fields: vec![
-                    Field::unnamed(DataType::Int64),
-                    Field::unnamed(DataType::Int64),
-                ],
-            };
-            Box::new(MockSource::with_messages(
-                schema,
-                pk_indices(),
-                vec![
-                    Message::Barrier(Barrier::new_test_barrier(3)),
-                    Message::Chunk(std::mem::take(&mut chunks[0])),
-                    Message::Chunk(std::mem::take(&mut chunks[1])),
-                    Message::Barrier(Barrier::new_test_barrier(4)),
-                ],
-            ))
-        }
-
-        #[tokio::test]
-        async fn test_with_ties_recovery() {
-            let state_store = MemoryStateStore::new();
-            let state_table = create_in_memory_state_table_from_state_store(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-                state_store.clone(),
-            )
-            .await;
-            let top_n_executor = Box::new(
-                TopNExecutor::new_with_ties_for_test(
-                    create_source_before_recovery() as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (0, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor.execute();
-
-            // consume the init barrier
-            top_n_executor.next().await.unwrap().unwrap();
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    + 1 0
-                    + 2 1
-                    + 3 2"
-                )
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    + 3 6
-                    + 3 7
-                    - 3 7
-                    - 3 6
-                    - 3 2
-                    + 1 8
-                    + 2 9"
-                )
-            );
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let state_table = create_in_memory_state_table_from_state_store(
-                &[DataType::Int64, DataType::Int64],
-                &[OrderType::ascending(), OrderType::ascending()],
-                &pk_indices(),
-                state_store,
-            )
-            .await;
-
-            // recovery
-            let top_n_executor_after_recovery = Box::new(
-                TopNExecutor::new_with_ties_for_test(
-                    create_source_after_recovery() as Box<dyn Executor>,
-                    ActorContext::create(0),
-                    storage_key(),
-                    (0, 3),
-                    order_by(),
-                    1,
-                    state_table,
-                )
-                .unwrap(),
-            );
-            let mut top_n_executor = top_n_executor_after_recovery.execute();
-
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 0"
-                )
-            );
-
-            // High cache has only 2 capacity, but we need to trigger 3 inserts here!
-            let res = top_n_executor.next().await.unwrap().unwrap();
-            assert_eq!(
-                *res.as_chunk().unwrap(),
-                StreamChunk::from_pretty(
-                    " I I
-                    - 1 8
-                    + 3 2
-                    + 3 6
-                    + 3 7
-                    "
-                )
-            );
-            println!("hello");
-            // barrier
-            assert_matches!(
-                top_n_executor.next().await.unwrap().unwrap(),
-                Message::Barrier(_)
-            );
-        }
-    }
-}
