+++ src/stream/tests/it/main.rs
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod prelude {
-    pub use risingwave_common::array::StreamChunk;
-    pub use risingwave_common::catalog::{Field, Schema};
-    pub use risingwave_common::test_prelude::StreamChunkTestExt;
-    pub use risingwave_common::types::DataType;
-    pub use risingwave_expr::expr::build_from_pretty;
-    pub use risingwave_expr::table_function::repeat;
-    pub use risingwave_stream::executor::test_utils::snapshot::*;
-    pub use risingwave_stream::executor::test_utils::{MessageSender, MockSource};
-    pub use risingwave_stream::executor::{BoxedMessageStream, Executor, PkIndices};
-}
+// tests
+mod eowc;
 mod project_set;
+
+// utils
+mod prelude {
+    pub use expect_test::{expect, expect_file};
+    pub use risingwave_stream::executor::test_utils::prelude::*;
+
+    pub use crate::snapshot::*;
+}
+mod snapshot;
